requirements/dev.txt
-e .
black==21.5b1
flake8==3.9.2
flake8-isort==4.0.0
mypy==0.812
pre-commit==2.12.1
pytest==6.2.4
pytest-aiohttp==0.3.0
pytest-asyncio==0.14.0
